PREDDIMER TMDOCK closedist calculation
Former-commit-id: 003caa38f1810948ed576f1beb7d0265f45e847f [formerly c93588b722d95a9492429c4febd7f8f915c50c6d] [formerly 8ac8bb9a3718a824bff20a838c7e0d01385ca369 [formerly e00482059900622c5141decf5e88612b7296ab17]]
Former-commit-id: c82ddb8b1b776b01d48c94e09088806ade4add08 [formerly 45bf351915204d520a6ef7a74ec94256fd6909ea]
Former-commit-id: f7336eb3d0f21c66ea18dc1d4d166583c28216ec

+++ thoipapy/NCBI_BLAST/download/download.py
@@ -131,12 +131,13 @@ def download_homologues_from_ncbi(acc, TMD_seq_pl_surr, blast_xml_file, xml_txt,
     with open(xml_txt, "w") as f:
         f.write("acc\t{}\ndownload_date\t{}\ndatabase\tncbi_nr\nexpect_value\t{}\n".format(acc, date, expect_value))
 
-    with tarfile.open(xml_tar_gz, mode='w:gz') as tar:
-        # add the files to the compressed tarfile
-        tar.add(blast_xml_file, arcname=os.path.basename(blast_xml_file))
-        tar.add(xml_txt, arcname=os.path.basename(xml_txt))
+    if os.path.isfile(blast_xml_file):
+        with tarfile.open(xml_tar_gz, mode='w:gz') as tar:
+            # add the files to the compressed tarfile
+            tar.add(blast_xml_file, arcname=os.path.basename(blast_xml_file))
+            tar.add(xml_txt, arcname=os.path.basename(xml_txt))
 
-    delete_BLAST_xml(blast_xml_file)
+        delete_BLAST_xml(blast_xml_file)
 
     logging.info("Output file: {}. (time taken = {:0.3f} min)".format(xml_tar_gz, duration / 60))
